wip(cachecomponents): extract ShowCard today-showtimes to a Client Component
